+++ server/src/main/java/org/hkurh/doky/password/TokenService.kt
@@ -1,0 +1,8 @@
+package org.hkurh.doky.password
+
+import java.util.*
+
+interface TokenService {
+    fun calculateExpirationDate(): Date
+    fun generateToken(): String
+}
