+++ server/src/main/java/org/hkurh/doky/password/impl/DefaultTokenService.kt
@@ -1,5 +1,6 @@
-package org.hkurh.doky.password
+package org.hkurh.doky.password.impl
 
+import org.hkurh.doky.password.TokenService
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import java.time.Instant
@@ -9,20 +10,20 @@ import java.time.ZoneOffset
 import java.util.*
 
 @Component
-class TokenUtil {
+class DefaultTokenService : TokenService {
 
     private val zoneId = ZoneId.of("UTC")
 
     @Value("\${doky.password.reset.token.duration}")
     var resetTokenDuration: Long = 10
 
-    fun calculateExpirationDate(): Date {
+    override fun calculateExpirationDate(): Date {
         val currentDate = LocalDateTime.ofInstant(Instant.now(), zoneId)
         val expiredDate = currentDate.plusMinutes(resetTokenDuration)
         return Date.from(expiredDate.toInstant(ZoneOffset.UTC))
     }
 
-    fun generateToken() : String {
+    override fun generateToken(): String {
         return UUID.randomUUID().toString()
     }
 }
